Bug 7270: fix types to reduce warnings of gcc

+++ sources/common-dylan/freebsd-common-extensions-helper.c
@@ -9,13 +9,13 @@ uint application_filename_length () {
   return len;
 }
 
-uint application_filename_name (char* buffer, uint len) {
+uint application_filename_name (char* buffer, size_t len) {
   int mib[4] = { CTL_KERN, KERN_PROC, KERN_PROC_PATHNAME, -1 };
   sysctl(mib, 4, buffer, &len, NULL, 0);
   return len;
 }
 
-uint application_arguments (char* buffer, uint len) {
+uint application_arguments (char* buffer, size_t len) {
   int mib[4] = { CTL_KERN, KERN_PROC, KERN_PROC_ARGS, getpid() };
   sysctl(mib, 4, buffer, &len, NULL, 0);
   return len;
